requirements.txt
fastapi
uvicorn
httpx
pydantic
python-multipart
jinja2
python-dotenv
